Move map code to separate component
Preparing for the stack navigator
